final cleanup before slp demo video
