+++ lib/pages/view_summoner.dart
@@ -1,14 +1,17 @@
 import 'package:dart_lol/lol_api/dart_lol.dart';
 import 'package:dart_lol/lol_api/summoner.dart';
 import 'package:flutter/material.dart';
+import 'package:kronic_desktop_tool/pages/home_view.dart';
 
 class ViewSummoner extends StatefulWidget {
   static const routeName = '/view_summoner';
+  String? summoner;
+  ViewSummoner();
+  ViewSummoner.fromPlayer(this.summoner);
   @override
   _ViewSummonerState createState() => _ViewSummonerState();
 }
 
-//TODO Make Rank Look Good
 class _ViewSummonerState extends State<ViewSummoner> {
   var version;
   Future<Summoner> setupSummoner(String value) async {
@@ -16,8 +19,7 @@ class _ViewSummonerState extends State<ViewSummoner> {
     Summoner instance = Summoner.fromName(value);
     await instance.setupSearchedSummoner();
     //Check to make sure Data is Valid
-    if (instance.accountId == null ||
-        instance.summonerName == null) {
+    if (instance.accountId == null || instance.summonerName == null) {
       throw StateError('Error');
     }
     return instance;
@@ -164,20 +166,18 @@ class _ViewSummonerState extends State<ViewSummoner> {
   }
 
   int getQueue(int soloOrFlex, Summoner data) {
-    if (data.rank.ranks.length == 0)
-      {
-        return -1;
-      }
+    if (data.rank.ranks.length == 0) {
+      return -1;
+    }
     if (soloOrFlex == 1) {
-
-        for (int i = 0; i <= data.rank.ranks.length - 1; i++) {
-            if (data.rank.ranks[i].queueType == "RANKED_SOLO_5x5") {
-              return i;
-            }
-          }
-        // No Solo Duo Rank
-        return -1;
-      } else {
+      for (int i = 0; i <= data.rank.ranks.length - 1; i++) {
+        if (data.rank.ranks[i].queueType == "RANKED_SOLO_5x5") {
+          return i;
+        }
+      }
+      // No Solo Duo Rank
+      return -1;
+    } else {
       for (int i = 0; i <= data.rank.ranks.length - 1; i++) {
         if (data.rank.ranks[i].queueType == "RANKED_FLEX_SR") {
           return i;
@@ -189,34 +189,32 @@ class _ViewSummonerState extends State<ViewSummoner> {
   }
 
   Widget _buildRankInfo(Summoner data, int i) {
-
     if (i == 0) {
       var currentQueue = getQueue(1, data);
       if (currentQueue == -1) {
-          return Text("Unranked");
-        }
-        // Solo Duo on Top
-        return Column(
-          children: <Widget>[
-            Text('Solo/Duo',
-                style: TextStyle(
-                    color: Colors.yellowAccent, fontWeight: FontWeight.bold)),
-
-            Text(data.rank.ranks[currentQueue].tier,
-                style: TextStyle(fontWeight: FontWeight.bold)),
-            Text(data.rank.ranks[currentQueue].rank,
-                style: TextStyle(fontWeight: FontWeight.bold)),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: <Widget>[
-                Text(data.rank.ranks[currentQueue].lp.toString(),
-                    style: TextStyle(fontWeight: FontWeight.bold)),
-                Text('LP', style: TextStyle(fontWeight: FontWeight.bold)),
-              ],
-            )
-          ],
-        );
-      } else {
+        return Text("Unranked");
+      }
+      // Solo Duo on Top
+      return Column(
+        children: <Widget>[
+          Text('Solo/Duo',
+              style: TextStyle(
+                  color: Colors.yellowAccent, fontWeight: FontWeight.bold)),
+          Text(data.rank.ranks[currentQueue].tier,
+              style: TextStyle(fontWeight: FontWeight.bold)),
+          Text(data.rank.ranks[currentQueue].rank,
+              style: TextStyle(fontWeight: FontWeight.bold)),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              Text(data.rank.ranks[currentQueue].lp.toString(),
+                  style: TextStyle(fontWeight: FontWeight.bold)),
+              Text('LP', style: TextStyle(fontWeight: FontWeight.bold)),
+            ],
+          )
+        ],
+      );
+    } else {
       // Flex on Bottom
       var currentQueue = getQueue(2, data);
       if (currentQueue != -1) {
@@ -247,63 +245,12 @@ class _ViewSummonerState extends State<ViewSummoner> {
                   color: Colors.yellowAccent, fontWeight: FontWeight.bold)),
           Text("Unranked",
               style: TextStyle(fontWeight: FontWeight.bold, fontSize: 25)),
-          SizedBox(height: 40,)
-            ],
-          );
+          SizedBox(
+            height: 40,
+          )
+        ],
+      );
     }
-
-  }
-
-  Widget _buildRank(Summoner data) {
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      mainAxisAlignment: MainAxisAlignment.start,
-      children: <Widget>[
-        Expanded(
-          child: Container(
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.all(Radius.circular(8)),
-              color: Color.fromRGBO(43, 38, 60, 1),
-            ),
-            //margin: EdgeInsets.fromLTRB(0, 0, 700, 0),
-            child: Column(
-              children: <Widget>[
-                Row(
-                  // Outer Top Row
-                  children: <Widget>[
-                    // Rank
-                    Expanded(
-                      child: Column (
-                        children: <Widget> [
-                          Container(
-                            decoration: BoxDecoration(
-                              border: Border.all(width: 2, color: Colors.grey.shade600)
-                            ),
-                            child: ListView.separated(
-                              separatorBuilder: (context, index) => Divider(
-                                color: Colors.grey,
-                              ),
-                              itemCount: 2,
-                              shrinkWrap: true,
-                              addAutomaticKeepAlives: true,
-                              itemBuilder: (context, i) {
-                                return _buildRankInfo(data, i);
-                              },
-                            ),
-                          )
-                        ]
-                      ),
-                    ),
-
-                    SizedBox(width: 800),
-                  ],
-                ),
-              ],
-            ),
-          ),
-        ),
-      ],
-    );
   }
 
   Widget _buildMatchHistory(Summoner data) {
@@ -318,29 +265,93 @@ class _ViewSummonerState extends State<ViewSummoner> {
         });
   }
 
+  Widget playerCard(Summoner summoner) {
+    double boxHeight = 0;
+    if (summoner.rank.ranks.length != 0) {
+      boxHeight = 150 + (summoner.rank.ranks.length * 80);
+    } else {
+      boxHeight = 150;
+    }
+    return Stack(
+      children: <Widget>[
+        Card(
+          margin: const EdgeInsets.only(top: 20.0),
+          child: SizedBox(
+              height: boxHeight,
+              width: 300,
+              child: Padding(
+                padding: const EdgeInsets.only(top: 45.0),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: <Widget>[
+                    Text(
+                      summoner.summonerName!,
+                      style: TextStyle(
+                        fontSize: 50,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    ),
+                    Text("Level: " + summoner.summonerLevel.toString()),
+                    ListView.separated(
+                      separatorBuilder: (context, index) => Divider(
+                        color: Colors.yellow,
+                      ),
+                      itemCount: summoner.rank.ranks.length,
+                      addAutomaticKeepAlives: true,
+                      shrinkWrap: true,
+                      itemBuilder: (context, i) {
+                        if (summoner.rank.ranks[i].tier == null) {
+                          return Container();
+                        }
+                        return _buildRankInfo(summoner, i);
+                      },
+                    ),
+                  ],
+                ),
+              )),
+        ),
+        Positioned(
+          top: .0,
+          left: .0,
+          right: .0,
+          child: Center(
+            child: FutureBuilder<NetworkImage>(
+                future: summonerIcon(summoner.profileIconId!),
+                builder: (context, snapshot) {
+                  if (snapshot.hasData) {
+                    return CircleAvatar(
+                      radius: 30,
+                      foregroundImage: snapshot.data,
+                    );
+                  }
+                  return CircleAvatar();
+                }),
+          ),
+        )
+      ],
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
-    String? data = ModalRoute.of(context)?.settings.arguments as String?;
+    String? data;
+    if (widget.summoner == null) {
+      data = ModalRoute.of(context)?.settings.arguments as String?;
+    } else {
+      data = widget.summoner;
+    }
 
     return FutureBuilder<Summoner>(
         future: setupSummoner(data!),
         builder: (BuildContext context, AsyncSnapshot<Summoner> snapshot) {
           if (snapshot.hasData) {
             return Scaffold(
-              backgroundColor: Color.fromRGBO(28, 22, 46, 1),
               appBar: AppBar(
-                title: Text(snapshot.data!.summonerName!,
-                    style: TextStyle(fontWeight: FontWeight.bold)),
-                elevation: 0.0,
-                backgroundColor: Color.fromRGBO(28, 22, 46, 1),
-              ),
-              body: Column(
+                  elevation: 0, backgroundColor: Color.fromRGBO(28, 22, 46, 1)),
+              backgroundColor: Color.fromRGBO(28, 22, 46, 1),
+              body: Row(
                 children: <Widget>[
-                  Divider(
-                    color: Color.fromRGBO(43, 38, 60, 1),
-                    thickness: 2,
-                  ),
-                  _buildRank(snapshot.data!),
+                  playerCard(snapshot.data!),
                   Expanded(
                     child: Container(
                       width: 500,
