+++ pubspec.yaml
@@ -59,6 +59,7 @@ dev_dependencies:
 flutter:
   assets:
     - assets/images/league_home.jpg
+    - assets/images/kaisa_corrupted_opt.jpg
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
